third_party/blink/web_tests/platform/generic/fast/forms/file/file-input-empty-validation-expected.txt

 

This tests that a file input cleared of its (required) value doesn't validate.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS invalid is true
PASS successfullyParsed is true

TEST COMPLETE
